refactor: extract TransientToastTypes.swift from TransientToast.swift
Extracts 2 support types (TransientToastStyle enum + TransientToastAction struct)
from TransientToast.swift into a new TransientToastTypes.swift.

Byte-identical move; SHA verified against origin/main.

Reduces TransientToast.swift from 52 to 20 lines. Only the primary
TransientToast struct remains.

Tests: 667 (unchanged).
